m/requirements.txt
# Django AI Chatbot Project Dependencies
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
duckduckgo-search==3.9.6
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.3
huggingface-hub==0.19.4
transformers==4.35.2
torch==2.1.1
numpy==1.24.3
pandas==2.1.3
pillow==10.1.0
django-cors-headers==4.3.1
django-celery-beat==2.5.0
channels==4.0.0
channels-redis==4.1.0
websockets==12.0
googletrans==4.0.0rc1
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-extensions==3.2.3
sqlparse==0.4.4
